+++ examples/basic_usage.py
@@ -7,8 +7,13 @@ client = PokeLance()
 
 async def main() -> None:
     print(await client.ping())
-    x = await client.pokemon.fetch_pokemon_form("charmander")
-    print(x.to_dict())
+    print(await client.berry.fetch_berry("cheri"))
+    print(await client.berry.fetch_berry_flavor("spicy"))
+    print(await client.berry.fetch_berry_firmness("very-soft"))
+    print(client.berry.get_berry("cheri"))
+    print(client.berry.get_berry_flavor("spicy"))
+    print(client.berry.get_berry_firmness("very-soft"))
+    await client.close()
     return None